# Faster by=key(DT)[1] and keyby= index