--- a/inst/tests/tests.Rraw
+++ b/inst/tests/tests.Rraw
@@ -12177,6 +12177,27 @@ test(1940, data.table(A=1:3, .SD=4:6), error=".SD.*has special meaning")
 test(1941.1, fintersect(data.table(y=1), data.table(y=2)), data.table(y=numeric()))
 test(1941.2, fintersect(data.table(y=1), data.table(y=1)), data.table(y=1))
 
+# by= and keyby= using uniqlist on key,  keyby= using index and passing it to uniqlist
+DT = data.table(id=c("D","A","C","A","C"), v=1:5)
+setkey(DT,id)
+test(1942.1, DT[,sum(v),by=id,verbose=TRUE], data.table(id=c("A","C","D"), V1=INT(6,8,1), key="id"), output="Finding groups using uniqlist on key")
+DT = data.table(id1=c("D","A","C","A","C"), id2=INT(3,9,2,9,3), id3=c(3.4, 9.1, 3.4, 3.3, 9.1), v=1:5)
+setindex(DT,id1)
+old = options(datatable.use.index=TRUE)
+test(1942.2, DT[,sum(v),keyby=id1,verbose=TRUE], data.table(id1=c("D","A","C"), V1=INT(1,6,8), key="id1"), output="Finding groups using uniqlist on index 'id1'")
+setindex(DT,id2)
+test(1942.3, DT[,sum(v),keyby=id2,verbose=TRUE], data.table(id2=INT(3,9,2), V1=INT(6,6,3), key="id2"), output="Finding groups using uniqlist on index 'id2")
+setindex(DT,id3)
+test(1942.4, DT[,sum(v),keyby=id3,verbose=TRUE], data.table(id3=c(3.4, 9.1, 3.3), V1=INT(4,7,4), key="id3"), output="Finding groups using uniqlist on index 'id3'")
+setindex(DT, NULL)
+test(1942.5, indices(DT), NULL)
+setindex(DT,id1,id2)
+test(1942.6, DT[,sum(v),keyby=id1,verbose=TRUE], data.table(id1=c("D","A","C"), V1=INT(1,6,8), key="id1"), output="Finding groups using uniqlist on index 'id1__id2'")
+test(1942.7, DT[,sum(v),keyby=.(id1,id2),verbose=TRUE], data.table(id1=c("A","C","C","D"), id2=INT(9,2,3,3), V1=INT(6,3,5,1), key="id1,id2"), output="Finding groups using uniqlist on index 'id1__id2'")
+test(1942.8, DT[,sum(v),keyby=.(id2,id1),verbose=TRUE], data.table(id2=INT(2,3,3,9), id1=c("C","C","D","A"), V1=INT(3,5,1,6), key="id2,id1"), output="Finding groups using forderv")
+options(datatable.use.index=FALSE)
+test(1942.9, DT[,sum(v),keyby=id1,verbose=TRUE], data.table(id1=c("D","A","C"), V1=INT(1,6,8), key="id1"), output="Finding groups using forderv")
+options(old)
 
 ###################################
 #  Add new tests above this line  #