```
N = 5e8
set.seed(1)
DT = data.table(
  id = sample(LETTERS[1:6], N, replace=TRUE),
  val = sample(10, N, replace=TRUE)
)
setkey(DT,id)
```

Before this PR : 
```
> system.time(DT[,sum(val),by=id,verbose=TRUE]) 
Detected that j uses these columns: val 
Finding groups using uniqlist ... 6.924sec 
Finding group sizes from the positions (can be avoided to save RAM) ... 0.002sec 
lapply optimization is on, j unchanged as 'sum(val)'
GForce optimized j to 'gsum(val)'
Making each group and running j (GForce TRUE) ... 3.156sec 
   user  system elapsed 
 10.567   0.601  11.168 
```

With this PR : 
```
> system.time(DT[,sum(val),by=id,verbose=TRUE]) 
Detected that j uses these columns: val 
Finding groups using uniqlist ... 0.373sec 
Finding group sizes from the positions (can be avoided to save RAM) ... 0.002sec 
lapply optimization is on, j unchanged as 'sum(val)'
GForce optimized j to 'gsum(val)'
Making each group and running j (GForce TRUE) ... 3.190sec 
   user  system elapsed 
  4.046   0.615   4.662 
```

The `uniqlist` part reduces from 6.9s to 0.4s.

ToDo:

- [x] ~add missing test of by=key(DT)[1] where that is integer64~  wasn't missing.
- [x] numeric rounding to pass test 1196.2

Found as a result of investigating and thanks to #2962.