@jangorecki 2018-09-09
do we want this in patch release? I think there are potentially other improvements we could do here, thus might make sense to postpone to 1.12.0