@jangorecki 2018-09-12
The other improvements I had in mind was #3042, and it will probably takes some time. Lets get this one merged for 1.11.6, they don't seems to be that much related.